Update the tollbooth dependency

+++ engine/handlers.go
@@ -429,7 +429,7 @@ func (ac *Config) RegisterHandlers(mux *http.ServeMux, handlePath, servedir stri
 	if ac.disableRateLimiting {
 		mux.HandleFunc(handlePath, allRequests)
 	} else {
-		limiter := tollbooth.NewLimiter(ac.limitRequests, nil)
+		limiter := tollbooth.NewLimiter(float64(ac.limitRequests), nil)
 		limiter.SetMessage(themes.MessagePage("Rate-limit exceeded", "<div style='color:red'>You have reached the maximum request limit.</div>", theme))
 		limiter.SetMessageContentType("text/html; charset=utf-8")
 		mux.Handle(handlePath, tollbooth.LimitFuncHandler(limiter, allRequests))
