+++ engine/luahandler.go
@@ -44,7 +44,7 @@ func (ac *Config) LoadLuaHandlerFunctions(L *lua.LState, filename string, mux *h
 		if ac.disableRateLimiting {
 			mux.HandleFunc(handlePath, wrappedHandleFunc)
 		} else {
-			limiter := tollbooth.NewLimiter(ac.limitRequests, nil)
+			limiter := tollbooth.NewLimiter(float64(ac.limitRequests), nil)
 			limiter.SetMessage(themes.MessagePage("Rate-limit exceeded", "<div style='color:red'>You have reached the maximum request limit.</div>", theme))
 			limiter.SetMessageContentType("text/html; charset=utf-8")
 			mux.Handle(handlePath, tollbooth.LimitFuncHandler(limiter, wrappedHandleFunc))
